Make some c++ python calls private. Changed StreamWriterSlave to StreamWriterPort Stable state for new python library development

+++ include/rogue/utilities/PeekPoke.h
@@ -45,10 +45,10 @@ namespace rogue {
             ~PeekPoke();
 
             //! Poke. Write to an address
-            void poke ( uint64_t address, uint32_t value );
+            void poke ( uint64_t address, uint64_t value );
 
             //! Peek. Read from an address
-            uint32_t peek ( uint64_t address );
+            uint64_t peek ( uint64_t address );
       };
 
       // Convienence
